store logged user's instructor status to header
Check when user logins if their student number is stored in any group
that belong to the last created configuration. This information is
stored into the token.

Use this change to define a middleware function to check if a logged
token is classified as an instructor.

These cahanges are to enable alternate backend requests for data used
to develop admin and instructor views for time logs.